refactor: made changes to the user services
1. renamed user service to auth.
2. renamed user controller to auth.
3. renamed user.services.d to auth.service.d.
4. renamed createUser schema to register user schema.
5. renamed registration route.
6. replaced imports in the respective places.
