+++ src/schemas/auth.schema.ts
@@ -1,6 +1,6 @@
 import {object,string,TypeOf} from "zod"
 
-export const createUserSchema = object({
+export const registerUserSchema = object({
 	body:object({
 		email:string({required_error:"Email is a required field"}).email({message:"Not a valid email address"}),
 		password:string({required_error:"password is a required field"}).min(8,{message:"password should atleast be 8 charachters  long"}),
@@ -15,5 +15,5 @@ export const userLoginSchema = object({
 	})
 })
 
-export type createUserInputType = TypeOf<typeof createUserSchema>["body"]
+export type registerUserInputType = TypeOf<typeof registerUserSchema>["body"]
 export type userLoginInputType = TypeOf<typeof userLoginSchema>["body"]
