+++ src/services/auth.service.d.ts
@@ -2,13 +2,12 @@
 type SuccessUserCreationResult = {
     success: true
     id:string
-    status:number
 }
 
 type FailedUserCreationResult = {
     success: false
     error: string
-    status:number
+    errorCode:string
 }
     
 export type UserCreationResult = SuccessUserCreationResult | FailedUserCreationResult
